# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "API support for a E-Commerce",
        "contact": {
            "email": "makinsuroju@gmail.com"
        },
        "license": {
            "name": "Apache License 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get all products in cart",
                "description": "Get all products in cart",
                "operationId": "2721c2a7f5993ff11ef640d1ca9a2c7c",
                "responses": {
                    "200": {
                        "description": "Success operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add new item cart",
                "operationId": "d31e6a9806855ee4121c195b0ceea8dc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "product_id",
                                    "product_price",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "title": "product_id",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_name": {
                                        "title": "product_name",
                                        "type": "string",
                                        "example": "Gucci Shirt"
                                    },
                                    "product_price": {
                                        "title": "product_price",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "title": "quantity",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cart": {
                                            "$ref": "#/components/schemas/Cart"
                                        },
                                        "status": {
                                            "title": "status",
                                            "type": "string",
                                            "default": "success"
                                        },
                                        "status_code": {
                                            "title": "status_code",
                                            "type": "number",
                                            "default": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/cart/update/id": {
            "patch": {
                "tags": [
                    "Cart"
                ],
                "summary": "update quantity in cart",
                "operationId": "2129e67715666f495206449569d9adcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "title": "quantity",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/cart/remove/{id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "remove an item from cart",
                "description": "remove an item from cart",
                "operationId": "5d5ec3bd3c28d14beecff21ee47f96f7",
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear all cart items associated with the current user.",
                "description": "clear all items cart",
                "operationId": "dbfc2e8c424538ad5364e0ab0246eb0f",
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cart": {
                "title": "Cart",
                "required": [
                    "user_id",
                    "product_name",
                    "product_id",
                    "quantity",
                    "price"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer",
                        "default": "1"
                    },
                    "user_id": {
                        "title": "user_id",
                        "type": "integer",
                        "default": "1"
                    },
                    "product_name": {
                        "title": "product_name",
                        "type": "string",
                        "default": "Gucci Shirt"
                    },
                    "product_id": {
                        "title": "product_id",
                        "type": "integer",
                        "default": "1"
                    },
                    "quantity": {
                        "title": "quantity",
                        "type": "integer",
                        "default": "1"
                    },
                    "price": {
                        "title": "price",
                        "type": "double",
                        "default": "199.99"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}